requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cryptography==3.4.6
Django==3.2
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.6.2
idna==2.10
jwcrypto==0.8
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.1.2
pipdeptree==2.0.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
